More recursion

+++ GameState.h
@@ -149,7 +149,7 @@ public:
     XYPos left_panel_offset;
     XYPos right_panel_offset;
     XYPos grid_offset;
-    int grid_zoom = 0;
+    double grid_zoom = 0;
     bool grid_dragging = false;
     XYPos grid_dragging_last_pos;
     XYPos scaled_grid_offset;
